+++ audionimbus/CHANGELOG.md
@@ -24,6 +24,8 @@
 - Fixed a segmentation fault when calling `BinauralEffect::tail` with an output buffer that does not have two channels.
 - Fixed a segmentation fault when applying an ambisonics encode effect on an input buffer does not have exactly one channel or an output buffer that does not have the correct number of channels for the ambisonics order.
 - Fixed a segmentation fault when calling `AmbisonicsEncodeEffect::tail` with an output buffer that does not have the correct number of channels for the ambisonics order.
+- Fixed a segmentation fault when applying an ambisonics decode effect on an input buffer that does not have the correct number of channels for the ambisonics order, or an output buffer that does not have the correct number of channels for the speaker layout (or two channels for binaural) specified when creating the effect.
+- Fixed a segmentation fault when calling `AmbisonicsDecodeEffect::tail` with an output buffer that does not have the correct number of channels for the speaker layout (or two channels for binaural) specified when creating the effect.
 
 ### Changed
 
@@ -52,6 +54,9 @@
 - `VirtualSurroundEffect::tail` now returns an `EffectError` when the output buffer does not have two channels.
 - `AmbisonicsEncodeEfect::apply` now returns an `EffectError` when the input buffer does not have exactly one channel or the output buffer that does not have the correct number of channels for the ambisonics order.
 - `AmbisonicsEncodeEffect::tail` now returns an `EffectError` when the output buffer that does not have the correct number of channels for the ambisonics order.
+- `AmbisonicsDecodeEffect::apply` now returns an `EffectError` when the input buffer does not have the correct number of channels for the ambisonics order, or the output buffer does not have the correct number of channels for the speaker layout (or two channels for binaural) specified when creating the effect.
+- `AmbisonicsDecodeEffect::tail` now returns an `EffectError` when the output buffer does not have the correct number of channels for the speaker layout (or two channels for binaural) specified when creating the effect.
+- `AmbisonicsDecodeEffectParams::binaural` field removed and added field `AmbisonicsDecodeEffectSettings::rendering` in order to validate output buffers in `AmbisonicsDecodeEffect::tail`.
 
 ### Added
 
@@ -67,6 +72,7 @@
 - Add `EffectError` for errors that can occur when applying audio effects.
 - `SteamAudioError` implements `PartialEq`, `Copy` and `Clone`.
 - Add `ChannelRequirement` to specify the channel count requirement for an audio buffer.
+- Add `Rendering` enum to choose between decoding ambisonics using binaural rendering or panning.
 
 ### Removed